aisearch/requirements.txt

pocketbase==0.9.2
python-dotenv==1.0.0
chromadb==0.4.14
fastapi==0.104.0
uvicorn==0.23.2